autoppia_iwa copy/modules/webs_demo/web_2_demo_angular_django_personal_management/backend/pythondjangocrud/apps/payroll/urls.py
from rest_framework.routers import DefaultRouter

from pythondjangocrud.apps.payroll.views import PayrollViewSet

app_name = "payroll"

router = DefaultRouter()
router.register(r'payroll', PayrollViewSet)

urlpatterns = router.urls
